feat: GlobalHeader 임시 UI 제거 및 디자인 수정 (#21)
